app/requirements/base.txt
redis==5.0.0
alembic==1.11.2
bcrypt==4.0.1
databases[asyncpg]==0.7.0
orjson==3.9.4
psycopg2-binary==2.9.7
python-jose==3.3.0
SQLAlchemy==1.4.42
httpx==0.24.1
requests==2.31.0
types-requests==2.31.0.20240218

pydantic[email]==2.1.1
pydantic-settings==2.0.3
fastapi==0.101.1
uvicorn[standard]==0.23.2

sentry-sdk==1.29.2
celery==5.3.1
flower==2.0.1

async-asgi-testclient==1.4.11
pytest-env==0.8.1
sqlalchemy2-stubs==0.0.2a35
types-python-jose==3.3.4.8
itsdangerous==2.1.2
authlib==1.2.1

python-multipart==0.0.6
aiofiles==23.2.1
types-aiofiles==23.2.0
fastapi-pagination==0.12.19
fastapi-filter==1.0.0
python-docx==1.1.0
lxml==4.9.3

# GOOGLE
google-auth==2.28.1
google-auth-oauthlib==1.0.0

# PYTZ
pytz==2023.3
types-pytz==2023.3.1.1

# AI
openai==1.13.3
langchain==0.1.9
langchain-core==0.1.29
langchain-openai==0.0.7
langchain-community==0.0.21
beautifulsoup4==4.12.3
unstructured==0.11.8
tiktoken==0.6.0
html2text==2024.2.25
nest-asyncio==1.6.0
playwright==1.41.2

PyPDF2==3.0.1
pdfminer==20191125
